Refactored to use libuv provided size functions and added assertions.
After experiencing some memory allocation bugs, I found an asymetry between
what was reported by sizeof() and what the static library's version of some
sizes were. It turned out to be a difference in some preprocessor definitions
causing structure size mismatches. So the code was refactored to rely on the
sizes reported by the static library, and asserts are added to the initLibrary
function to help catch problems early on.